tasks: kernel: Honor prebuilt kernel flag
For devices that want to use a prebuilt kernel, TARGET_KERNEL_SOURCE
would still be set to TARGET_AUTO_KDIR, meaning the build system would
still try to build the kernel if TARGET_AUTO_KDIR was present.

Setting TARGET_PREBUILT_KERNEL indicates this is not wanted, so don't
attempt to do it.

Change-Id: Ic79b3ac1b9c946fd258ada43dce2b08bb74ea0d9
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>